requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorzero==2.0
gpiozero==1.6.2
idna==2.10
numpy==1.20.3
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
pygame==2.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
